All queries modified
Modified all the queries to request only the necessary data. We only need a single value for most of the displays. The queries were modified to request the last value in each table instead of requesting all the values, which takes a lot of time. This helps a lot with the reload time.