Create sw.json
This flow allows you to read up to 100 registers at a time, parse them into variables, and creates an object in msg.payload that the influx-batch can read and write to the db in a single query. The registers are for Schneider SW devices (SW4048 in my case). Some of them (voltages) are not returning logical values...